ink_runtime::api::assets_api

Trait AssetsAPI

pub trait AssetsAPI<T, I = Instance1>
where
    T: RuntimeEnv,
    T::Runtime: Config<I>,
    I: 'static,
{
    // Required methods
    fn create(
        &mut self,
        id: &<T::Runtime as Config<I>>::AssetId,
        owner: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        min_balance: <T::Runtime as Config<I>>::Balance,
    ) -> Result<(), DispatchError>;
    fn set_metadata(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        owner: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        name: Vec<u8>,
        symbol: Vec<u8>,
        decimals: u8,
    ) -> Result<(), DispatchError>;
    fn metadata(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
    ) -> (Vec<u8>, Vec<u8>, u8);
    fn approve(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        owner: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        delegate: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        amount: <T::Runtime as Config<I>>::Balance,
    ) -> Result<(), DispatchError>;
    fn mint_into(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        account: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        value: <T::Runtime as Config<I>>::Balance,
    ) -> Result<<T::Runtime as Config<I>>::Balance, DispatchError>;
    fn transfer(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        source: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        dest: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        amount: <T::Runtime as Config<I>>::Balance,
    ) -> Result<(), DispatchError>;
    fn balance_of(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        owner: impl IntoAccountId<AccountIdFor<T::Runtime>>,
    ) -> <T::Runtime as Config<I>>::Balance;
    fn total_supply(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
    ) -> <T::Runtime as Config<I>>::Balance;
    fn allowance(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
        owner: impl IntoAccountId<AccountIdFor<T::Runtime>>,
        delegate: impl IntoAccountId<AccountIdFor<T::Runtime>>,
    ) -> <T::Runtime as Config<I>>::Balance;
    fn asset_exists(
        &mut self,
        asset: &<T::Runtime as Config<I>>::AssetId,
    ) -> bool;
}

Assets API for the runtime.

Provides methods to create, mint, and manage assets in pallet-assets.

Required Methods

fn create( &mut self, id: &<T::Runtime as Config<I>>::AssetId, owner: impl IntoAccountId<AccountIdFor<T::Runtime>>, min_balance: <T::Runtime as Config<I>>::Balance, ) -> Result<(), DispatchError>

Creates value amount of tokens and assigns them to account, increasing the total supply.

Arguments

id - ID of the new asset to be created.

owner - The owner of the created asset (accepts any type convertible to AccountId).

min_balance - The asset amount one account need at least.

fn set_metadata( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, owner: impl IntoAccountId<AccountIdFor<T::Runtime>>, name: Vec<u8>, symbol: Vec<u8>, decimals: u8, ) -> Result<(), DispatchError>

Sets the metadata for an existing fungible asset.

Arguments

asset - ID of the asset.

owner - The owner of the asset (accepts any type convertible to AccountId).

name - Token name.

symbol - Token symbol.

decimals - Token decimals.

fn metadata( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, ) -> (Vec<u8>, Vec<u8>, u8)

Returns the metadata for an asset.

Arguments

asset - ID of the asset.

Returns

A tuple of (name, symbol, decimals).

fn approve( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, owner: impl IntoAccountId<AccountIdFor<T::Runtime>>, delegate: impl IntoAccountId<AccountIdFor<T::Runtime>>, amount: <T::Runtime as Config<I>>::Balance, ) -> Result<(), DispatchError>

Approves spender to spend value amount of tokens on behalf of the caller.

Successive calls of this method overwrite previous values.

Arguments

asset - ID of the asset.

owner - The account that owns the tokens (accepts any type convertible to AccountId).

delegate - The account that is allowed to spend the tokens (accepts any type convertible to AccountId).

amount - The number of tokens to approve.

fn mint_into( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, account: impl IntoAccountId<AccountIdFor<T::Runtime>>, value: <T::Runtime as Config<I>>::Balance, ) -> Result<<T::Runtime as Config<I>>::Balance, DispatchError>

Creates value amount of tokens and assigns them to account, increasing the total supply.

Arguments

asset - ID of the asset.

account - The account to be credited with the created tokens (accepts any type convertible to AccountId).

value - The number of tokens to mint.

fn transfer( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, source: impl IntoAccountId<AccountIdFor<T::Runtime>>, dest: impl IntoAccountId<AccountIdFor<T::Runtime>>, amount: <T::Runtime as Config<I>>::Balance, ) -> Result<(), DispatchError>

Transfer amount of tokens from origin to dest.

Arguments

asset - ID of the asset.

source - The account from which tokens are transferred (accepts any type convertible to AccountId).

dest - The account to which tokens are transferred (accepts any type convertible to AccountId).

amount - The number of tokens to transfer.

fn balance_of( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, owner: impl IntoAccountId<AccountIdFor<T::Runtime>>, ) -> <T::Runtime as Config<I>>::Balance

Returns the account balance for the specified owner.

Arguments

owner - The account whose balance is being queried (accepts any type convertible to AccountId).

fn total_supply( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, ) -> <T::Runtime as Config<I>>::Balance

Returns the total supply of the asset.

Arguments

asset - ID of the asset.

fn allowance( &mut self, asset: &<T::Runtime as Config<I>>::AssetId, owner: impl IntoAccountId<AccountIdFor<T::Runtime>>, delegate: impl IntoAccountId<AccountIdFor<T::Runtime>>, ) -> <T::Runtime as Config<I>>::Balance

Returns the allowance for a spender approved by an owner.

Arguments

asset - ID of the asset.

owner - The account that owns the tokens (accepts any type convertible to AccountId).

delegate - The account that is allowed to spend the tokens (accepts any type convertible to AccountId).

fn asset_exists(&mut self, asset: &<T::Runtime as Config<I>>::AssetId) -> bool

Check if the asset exists.

Arguments

asset - ID of the asset.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.