ink_runtime

Module precompiles

Export pallets that are used in crate::create_runtime

Modules

erc20

ERC-20 precompile interface for pallet-assets.

Functions

fixed_address

Calculates the address of a precompile at index n.

prefixed_address

Calculates the address of a precompile at index n with an additional prefix.