ink_runtime::precompiles::erc20

Trait Erc20

pub trait Erc20: ContractEnv {
    type totalSupplyOutput: ImpliesReturn<U256>;
    type balanceOfOutput: ImpliesReturn<U256>;
    type transferOutput: ImpliesReturn<bool>;
    type allowanceOutput: ImpliesReturn<U256>;
    type approveOutput: ImpliesReturn<bool>;
    type transferFromOutput: ImpliesReturn<bool>;

    // Required methods
    fn totalSupply(&self) -> Self::totalSupplyOutput;
    fn balanceOf(&self, account: H160) -> Self::balanceOfOutput;
    fn transfer(&mut self, to: H160, value: U256) -> Self::transferOutput;
    fn allowance(&self, owner: H160, spender: H160) -> Self::allowanceOutput;
    fn approve(&mut self, spender: H160, value: U256) -> Self::approveOutput;
    fn transferFrom(
        &mut self,
        from: H160,
        to: H160,
        value: U256,
    ) -> Self::transferFromOutput;
}

Defines the ERC-20 interface of the Asset Hub precompile.

Required Associated Types

type totalSupplyOutput: ImpliesReturn<U256>

Output type of the respective trait message.

type balanceOfOutput: ImpliesReturn<U256>

Output type of the respective trait message.

type transferOutput: ImpliesReturn<bool>

Output type of the respective trait message.

type allowanceOutput: ImpliesReturn<U256>

Output type of the respective trait message.

type approveOutput: ImpliesReturn<bool>

Output type of the respective trait message.

type transferFromOutput: ImpliesReturn<bool>

Output type of the respective trait message.

Required Methods

fn totalSupply(&self) -> Self::totalSupplyOutput

Returns the total supply of tokens.

Solidity Signature

function totalSupply() external view returns (uint256);

fn balanceOf(&self, account: H160) -> Self::balanceOfOutput

Returns the balance of an account.

Arguments

account - The address to query the balance of

Solidity Signature

function balanceOf(address account) external view returns (uint256);

fn transfer(&mut self, to: H160, value: U256) -> Self::transferOutput

Transfers tokens to another account.

Arguments

to - The recipient address

value - The amount of tokens to transfer

Returns

Returns true if the transfer was successful.

Solidity Signature

function transfer(address to, uint256 value) external returns (bool);

fn allowance(&self, owner: H160, spender: H160) -> Self::allowanceOutput

Returns the allowance for a spender on behalf of an owner.

This shows how many tokens spender is allowed to spend on behalf of owner.

Arguments

owner - The token owner’s address

spender - The spender’s address

Solidity Signature

function allowance(address owner, address spender) external view returns (uint256);

fn approve(&mut self, spender: H160, value: U256) -> Self::approveOutput

Approves a spender to spend tokens on behalf of the caller.

Arguments

spender - The address authorized to spend tokens

value - The maximum amount the spender can spend

Returns

Returns true if the approval was successful.

Solidity Signature

function approve(address spender, uint256 value) external returns (bool);

fn transferFrom( &mut self, from: H160, to: H160, value: U256, ) -> Self::transferFromOutput

Transfers tokens from one account to another using allowance.

The caller must have sufficient allowance from the from account.

Arguments

from - The address to transfer tokens from

to - The recipient address

value - The amount of tokens to transfer

Returns

Returns true if the transfer was successful.

Solidity Signature

function transferFrom(address from, address to, uint256 value) external returns (bool);